requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
e2b>=0.13.0
eventlet==0.33.3
exceptiongroup==1.3.0
fastapi==0.115.12
gunicorn==21.2.0
h11==0.16.0
idna==3.10
matplotlib>=3.8.0
numpy>=1.24.0
openai>=1.0.0
pandas>=2.1.0
plotly>=5.18.0
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
python-dotenv>=1.0.0
python-engineio==4.7.1
python-pptx
python-socketio==5.9.0
requests>=2.31.0
sarvamai==0.1.0
seaborn>=0.13.0
sniffio==1.3.1
starlette==0.46.2
streamlit>=1.30.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
websockets==12.0
yt-dlp==2023.10.13
